langchain/chains/parallel/requirements.txt
langchain-groq==0.1.9
streamlit==1.37.0
python-dotenv>=1.0.1
langchain>=0.2.1